Handle All Warnings (#2)
* set build options

* remove COMPARE and MDOERN_* switches

* remove tools makefile

* AR patching is gone too since we want a fullly decomped version

* AR is modern

* remove cwd changes

* edit my own tool to fix compile errors

* compile files generated with my own tool instead of the originals

* inline modern_gcc makefile

* port mips toolchain detection logic

* add util.mk for find-command

* remove forced AR order and strip/mdebug removal commands

* add -mabi=32 to as flags

* cseq warning

* csplayer

* env

* voicecheckword

* voicestopreaddata

* voicedata

* voicecontwrite4

* voice done

* sinf

* sched

* rg

* cosf

* timerintr

* settimer

* xprintf

* xldtob

* contram

* conteepwrite

* conteepread

* dumpturbo

* gt

* aisetnextbuf

* us2dex_emu

* loadtextureblockmipmap

* guloadtile_bug

* kmcprintf

* seq

* sndplayer

* synthesizer

* seqplayer again

* load

* fix warnings

* formatting

* address csplayer, gu

* remove if(0)

* another explicit fallthrough

* last explicit fallthrough

* simplify triple multiply

* its 4 multiplies

* its an octuple multiply

* CI formatting

* sltu in alRaw16Pull

* another sltu

* handle all sltu's

* sndplayer sltu

---------

+++ src/audio/seq.c
@@ -240,8 +240,7 @@ void alSeqNewMarker(ALSeq* seq, ALSeqMarker* m, u32 ticks) {
                 lastTicks = seq->lastTicks;
                 break;
             }
-
-        } while (seq->lastTicks < ticks);
+        } while ((u32)seq->lastTicks < ticks);
 
         m->curPtr = lastPtr;
         m->lastStatus = lastStatus;
